# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScriptHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scripts:
    get:
      operationId: listScripts
      tags: [scripts]
      summary: List all scripts owned by current user
      responses:
        "200":
          description: List of scripts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Script"
    post:
      operationId: createScript
      tags: [scripts]
      summary: Create a new script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptInput"
      responses:
        "201":
          description: Created script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
  /scripts/{id}:
    get:
      operationId: getScript
      tags: [scripts]
      summary: Get a script by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateScript
      tags: [scripts]
      summary: Update a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptUpdate"
      responses:
        "200":
          description: Updated script
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
    delete:
      operationId: deleteScript
      tags: [scripts]
      summary: Delete a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /scripts/{id}/toggle:
    post:
      operationId: toggleScript
      tags: [scripts]
      summary: Enable or disable a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated script status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Script"
...
  /servers/{id}:
    delete:
      operationId: disconnectServer
      tags: [servers]
      summary: Disconnect a Discord server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /scripts/{id}/whitelist:
    get:
      operationId: listWhitelist
      tags: [whitelist]
      summary: List whitelist entries for a script
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of whitelist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhitelistEntry"
    post:
      operationId: addToWhitelist
      tags: [whitelist]
      summary: Add a Discord user to script whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhitelistInput"
      responses:
        "201":
          description: Added whitelist entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
  /scripts/{id}/whitelist/{entryId}:
    delete:
      operationId: removeFromWhitelist
      tags: [whitelist]
      summary: Remove a Discord user from script whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Script:
      type: object
      required: [id, ownerId, name, version, status, createdAt]
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        version:
          type: string
        status:
          type: string
          enum: [active, disabled]
        createdAt:
          type: string
    ScriptInput:
      type: object
      required: [name, version]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: string
          minLength: 1
    ScriptUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        version:
          type: string
          minLength: 1
        status:
          type: string
          enum: [active, disabled]
...
    WhitelistEntry:
      type: object
      required: [id, scriptId, discordUserId, createdAt]
      properties:
        id:
          type: integer
        scriptId:
          type: integer
        discordUserId:
          type: string
        addedBy:
          type: ["string", "null"]
        createdAt:
          type: string
    WhitelistInput:
      type: object
      required: [discordUserId]
      properties:
        discordUserId:
          type: string
        addedBy:
          type: string